[BRC-2431] Add end to end proxy protocol integration test for postgres filter (#289)
Add end to end integration with proxy protocol and verify that the original source ip address is sent to the sidecar external authz call.

Proxy protocol is used in AWS front end private link. There is no quick way to test that we can pass the real ip addr to the sidecar easily so I added integration test for this.

This PR added a new suite that will have proxy protocol listener before postgres_inspector. The test will then verify the source ip sent to the sidecar request. It also verifies the real ip addr output in the access log.