network: populate NNS in the filter state when connection is accepted (#41799)
This PR adds a change to start populating the Network Namespace as a
filter state value when the NNS is configured and connection gets
accepted. This helps using this value in filter state matcher or in
access logs.

---

**Commit Message:** network: populate NNS in the filter state when
connection is accepted
**Additional Description:** Added a change to start populating the
filter state `envoy.network.network_namespace` when a connection is
accepted on a listener.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>